Add tests
- Used minitest since it comes with ruby now, and was easy to work with
- Uses `rugged` to help with git validation
- There is a large `TestRepo` test helper class that was added

The `TestRepo` is probably the elephant in the room, since there is a
lot of code being added for it here, but this was intentional to help
facilitate adding new features in the future while allowing some
coverage for what already exists.

As mentioned in the comments, it was lifted from this library in
manageiq:

  https://github.com/NickLaMuro/manageiq/blob/master/spec/support/fake_ansible_repo.rb#L1

Part of this commit:

  https://github.com/NickLaMuro/manageiq/commit/652105e4

I have added some extra DSL methods an functionality not found in the
original, and removed all of the "Ansible bits", so it is a bit of a
departure from that code, but the core interface is the same.  I think
it makes writing tests for this project relatively easy.

In addition, there is a `CodeExtractor::TestCase` class in that same
`test_helper.rb` file which handles creating a sandbox for each test
method.  You can also run the tests with `DEBUG=1 rake` to avoid
deleting the directories after each run, which is nice for debugging
your tests when it isn't doing what you expect.