[CodeExtractor] Convert to module
This could have been done as part of the "Gemify" effort, but the intent
of that branch was to make as few changes to the core script before
starting to make sweeping changes.

By converting `CodeExtractor` to a `module`, that file and
`lib/code_extractor/version.rb` can be loaded at the same time (yes,
they couldn't before...), and it is just a better practice for dealing
with namespaces in general.

The `Runner` class was a necessary side effect, but I think it is mostly
understandable what it's role is and should be a straight forward
change.

Fun Fact
--------

The `attr_reader :extraction` is actually never used.  Kept it as a part
of `CodeExtractor::Runner`, but really, it is never used.