[CodeExtractor] Break out GitProject
This commit separates the concerns of the runner and the GitProject to
define methods that make sense for a generic GitProject.

In the future, the intent is that this lib should be able to move code
between two separate and already existing `GitProject` instances, and
having this abstraction layer in place will allow code to be shared more
easily.

This also still attempts to retain the "script-like" nature of the
original project in the `Runner`, but hopefully creating this generic
`GitProject` construct will facilitate future improvements.