[CodeExtractor] Add reinsert functionality
This is effectively the "undo" feature of the original project.

By that, it allows a user to take a portion (or all) of a previously
extracted project, and re-insert it back into the original project from
which it came from.

This becomes tricky, as we need to filter out the existing commits that
existed previously in the old ("target") repository.  As a result, 3
`filter-branch` passes are done to achieve this change:

1. After determining all file names that existed for the current
extractions, filter commits to down to ones that only include the files
we wish to commit.  In addition, create a script to move any of those
files into a temporary directory, maintaining the directory structure as
we go.
2. Take the temporary directory that was created and do a second filter
that makes the new root of the project the temporary directory we
created in the previous filter.
3. After adding the target repository as a remote, filter the commits
out that already exist in that remote.  Also in this step, a commit is
re-written that is shared between both repos that will be the
"re-inject" commit, which all of the injected commits will be based off
of.

From there, cherry-pick commits onto a new branch that is based off the
existing HEAD of the current target remote's branch (most likely
`master`) that will be receiving the "injected" commits.  This allows
the target repo to have the commits that were created "post extraction"
that are based off a commit that re-adds the code in a state where it
was originally extracted from.