[code_extractor_test] Add .perform_merges_of_extracted_code
Another helper method for breaking up some of the setup code for the
"unextract" test, which handles the portion of the setup that replicates
the process of merging the original extraction.

Not the best implementation, since `@bare_repo_dir` gets set as part of
the new method, and is used later in the test itself as the
`target_remote` in the `extraction_hash`, but for now it is a quick
solution.