[test_helper.rb] Add TestRepo.merge
Adds a helper function for handling merges via rugged.

Of note:  A `.checkout_head` call was added in addition to what was
there previously.  After doing a `commit` in rugged with a merge, the
git tree is left in a odd state (probably since the merge functionality
isn't "porcelain"), so this makes sure the git tree state is on HEAD.