[CodeExtractor] Support merge commits
When using cherry-pick, it wouldn't handle merges properly, since it
requires a `-m 1` argument for it to work properly.

By switching to using `git rebase --root --onto`, we can take the branch
that now is a "orphan" branch, and apply the commits onto the target
remote's target branch.

Left extra code in there to show other possibilities (since this isn't
getting merged in away)