[CodeExtractor] Use git user for base commit
When updating the commit that will be the root commit for the injected
commits, update the committer and author values so that it is more
obvious who re-added the commit.

As a result, also include the original committer/author info in the
updated commit message.  Formatting has been tweaked in this commit as
well to make it a little easier to see what is being done.