[FIXUP] Remove inject's --index-filter
This was a mistake to have this.  Since the commits will work just fine
as they are, and don't need to have the content of master to do so, this
can be removed entirely.

This step is much simpler now, as it is just updating the commits we are
bringing over and then applying them with a `git rebase --onto ...`.