[CodeExtractor] Fix "transferred from" in reinsert
This could be a "[FIXUP]" commit, but I think it is worth pointing out
what I did here.

Since there is multiple `git-filter-branch` calls, the original commit
SHAs that existed are re-written after each.  So delaying writing the
updated commit message will only produce references to commits that
don't exist since they were temporary.

So moving `--msg-filter` to the first `git-filter-branch` was the first
logical step.  However, this caused another issue.  Since we allow a
fallback for finding the commit messages that existed in the previous
repository, that would now be incorrect (in it's previous form) since it
relied on matching the exact messages created.

To handle this in most cases, and lines of `(transferred from ...@` are
filtered out from the commit messages before comparing.

* * *

The remaining changes are just to support this change.

  - Adjust some method signatures to support this change.
  - Update `assert_commits` to confirm the commit messages are pointing
    to a proper commit
  - Add a `reference_repo_dir` and `reference_repo` to help with the
    `assert_commits` changes (set in one of the helper methods)