[CodeExtractor] checkout before build_prune_script
Prior, when doing `build_prune_script`, the branch was left at what ever
state the previous step had left it at.  However, this caused and issue
when running `build_prune_script` as it was attempting to find the
history for files that were just deleted in the `extract_branch` step
(the branch that will be pushed to the `upstream_repo` that "deletes
code").

This step ensures a `git checkout #{source_branch}` prior to building
the script to ensure the files and their previous versions still exist.

And additional tweak was also added to split up the initialization of
the `file_and_ancestors` variable to make it easier to debug.