[WIP] grafting commits
This is a test to also include grafting as a feature (currently
disabled).

The intent is that a `graft_parent_commit` could be provided in the
config yaml, and that would be used in the changes here.

The steps in exampled in order:

1. Create a new branch with a HEAD at the "graft" point
2. Get back on the "inject_branch" (possibly unnecessary)
3. Rebase ("graft") the inject_branch onto the "old_branch"
4. Update the now grafted inject_branch to have the latest upstream
   changes

Step two is only "possibly unnecessary" if making use of another argument
of `git rebase` which I have yet to test.

Besides that though, the changes are pretty minimal to work with, so
this ends up being a pretty simple section to wrap your head around
(once you wrap your head around getting to this point in the first
place...).

Of note:  Since `inject_branch` is a branch with no parent commit for
it's root commit, the `--root` option is used here for `git rebase`,
which is important since the second arg is then the tip of
`inject_branch` (the branch name), and not the base of the branch that
you might normally provide when using `git rebase`.  This is important
since providing the base commit otherwise would "drop" that commit as
part of the rebase, and cause some other issues as a result (merge
conflicts).